+++ src/Mod/Drawing/App/ProjectionAlgos.h
@@ -39,7 +39,7 @@ class DrawingExport ProjectionAlgos
 {
 public:
     /// Constructor
-    ProjectionAlgos(const TopoDS_Shape &Input,const Base::Vector3f &Dir);
+    ProjectionAlgos(const TopoDS_Shape &Input,const Base::Vector3d &Dir);
     virtual ~ProjectionAlgos();
 
     void execute(void);
@@ -56,7 +56,7 @@ public:
 
 
     const TopoDS_Shape &Input;
-    const Base::Vector3f &Direction;
+    const Base::Vector3d &Direction;
 
     TopoDS_Shape V ;// hard edge visibly
     TopoDS_Shape V1;// Smoth edges visibly
